Updated trait files to new naming scheme
Note that, with the exception of filling in placeholder traits, all new
traits should be added in a new file with a higher number than the
previous last file. (So if it was 109~, the new file should be 110~) The
exception to this would be if your mod already has the highest numbered
file, in which case you just add traits at the end of that file. This
does require coordination with other HIP mods which add traits.